+++ src/atom.js
@@ -10,62 +10,62 @@ export const searchState = atom({
   default: "",
 });
 
-export const searchSelector = selector({
-  key: "searchSelector",
-  get: async ({ get }) => {
-    const searchWord = get(searchState);
-    // 디폴트 지역 -- 서울
-    if (!searchWord) {
-      try {
-        const fakeSeoulData = {
-          ok: true,
-          json: async () => ({
-            temp: 20,
-            date: "10-08",
-            weatherIcon: "02d",
-            name: "서울",
-          }),
-        };
+// export const searchSelector = selector({
+//   key: "searchSelector",
+//   get: async ({ get }) => {
+//     const searchWord = get(searchState);
+//     // 디폴트 지역 -- 서울
+//     if (!searchWord) {
+//       try {
+//         const fakeSeoulData = {
+//           ok: true,
+//           json: async () => ({
+//             temp: 20,
+//             date: "10-08",
+//             weatherIcon: "02d",
+//             name: "서울",
+//           }),
+//         };
 
-        const response = fakeSeoulData;
+//         const response = fakeSeoulData;
 
-        if (!response.ok) {
-          console.log("fail");
-          return null;
-        }
-        const data = await response.json();
-        return data;
-      } catch (error) {
-        console.error(error);
-        return null;
-      }
-    } else {
-      // 사용자 검색 지역
-      try {
-        const fakeData = {
-          ok: true,
-          json: async () => ({
-            temp: 15,
-            date: "10-09",
-            weatherIcon: "10n",
-            name: "전주",
-          }),
-        };
-        const response = fakeData;
+//         if (!response.ok) {
+//           console.log("fail");
+//           return null;
+//         }
+//         const data = await response.json();
+//         return data;
+//       } catch (error) {
+//         console.error(error);
+//         return null;
+//       }
+//     } else {
+//       // 사용자 검색 지역
+//       try {
+//         const fakeData = {
+//           ok: true,
+//           json: async () => ({
+//             temp: 15,
+//             date: "10-09",
+//             weatherIcon: "10n",
+//             name: "전주",
+//           }),
+//         };
+//         const response = fakeData;
 
-        if (!response.ok) {
-          console.log("fail");
-          return null;
-        }
-        const data = await response.json();
-        return data;
-      } catch (error) {
-        console.error(error);
-        return null;
-      }
-    }
-  },
-});
+//         if (!response.ok) {
+//           console.log("fail");
+//           return null;
+//         }
+//         const data = await response.json();
+//         return data;
+//       } catch (error) {
+//         console.error(error);
+//         return null;
+//       }
+//     }
+//   },
+// });
 
 // const response = await fetch("http://localhost:3000/api/weather", {
 //     method: "POST",
